requirements.txt
anthropic>=0.40.0,<1.0.0
chromadb==0.5.23
tavily-python>=0.3.0,<1.0.0
streamlit>=1.40.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
tiktoken>=0.7.0,<1.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
pytest>=8.0.0,<9.0.0
